Harden core runtime edge cases
