+++ .changeset/harden-core-runtime.md
@@ -1,0 +1,5 @@
+---
+"@anvia/core": patch
+---
+
+Harden MCP connection cleanup and vector dimension validation, and organize loader internals without changing public loader APIs.
