+++ packages/core/src/mcp/connect.ts
@@ -3,7 +3,17 @@ import type { McpConnection, McpServer } from "./types";
 
 export async function connectMcp(connection: McpConnection): Promise<McpServer> {
   const client = await connection.connect();
-  const { tools } = await client.listTools();
+  let tools: Awaited<ReturnType<typeof client.listTools>>["tools"];
+  try {
+    ({ tools } = await client.listTools());
+  } catch (error) {
+    try {
+      await client.close();
+    } catch {
+      // Preserve the initialization failure so callers see the actionable MCP error.
+    }
+    throw error;
+  }
 
   return {
     name: connection.name,
